api rename create to register add admin_clients add unregister api

+++ class/Gini/Controller/API/Wechat.php
@@ -8,8 +8,12 @@ class Wechat extends API {
 
     public function actionAuthorize($clientId, $clientSecret) {
         $clients = (array) \Gini\Config::get('app.clients');
+        $admin_clients = (array) \Gini\Config::get('app.admin_clients');
         if (isset($clients[$clientId]) && $clients[$clientId] == $clientSecret) {
             $_SESSION['app.client_id'] = $clientId;
+            if (in_array($clientId, $admin_clients)) {
+                $_SESSION['app.admin_client_id'] = $clientId;
+            }
             return session_id();
         }
         return false;
@@ -17,6 +21,10 @@ class Wechat extends API {
 
     private function isAuthorized() {
         return isset($_SESSION['app.client_id']);
+    }
+
+    private function isAdmin() {
+        return isset($_SESSION['app.admin_client_id']);
     }
 
     public function actionGetUnionId($token) {
@@ -62,9 +70,8 @@ class Wechat extends API {
         return $app->sendTemplateMessage($openId, $templateId, $data);
     }
 
-    public function actionCreateAppClient($clientId, $clientSecret) {
-        if (!$this->isAuthorized()) return false;
-
+    public function actionRegisterClient($clientId, $clientSecret) {
+        if (!$this->isAuthorized() || !$this->isAdmin()) return false;
         $confs     = \Gini\Config::Get('app');
         $env       = $_SERVER['GINI_ENV'];
         $base_path = APP_PATH.'/'.RAW_DIR.'/config/';
@@ -77,6 +84,32 @@ class Wechat extends API {
         }
         $confs['clients'][$clientId] = $clientSecret;
         $yaml_content                = yaml_emit($confs);
+        file_put_contents($file, $yaml_content);
+        \Gini\App\Cache::setup($env);
+        $new_confs = \Gini\Config::Get('app');
+        if ($new_confs == $confs) {
+            return true;
+        }
+        else {
+            return false;
+        }
+    }
+
+    public function actionUnregisterClient($clientId, $clientSecret) {
+        if (!$this->isAuthorized() || !$this->isAdmin()) return false;
+        $confs     = \Gini\Config::Get('app');
+        $env       = $_SERVER['GINI_ENV'];
+        $base_path = APP_PATH.'/'.RAW_DIR.'/config/';
+        $file      = $base_path.'@'.$env.'/app.yml';
+        if (!file_exists($file)) {
+            $file = $base_path.'app.yml';
+        }
+        if (!array_key_exists($clientId, (array)$confs['clients'])
+            || $confs['clients'][$clientId] != $clientSecret) {
+            return false;
+        }
+        unset($confs['clients'][$clientId]);
+        $yaml_content = yaml_emit($confs);
         file_put_contents($file, $yaml_content);
         \Gini\App\Cache::setup($env);
         $new_confs = \Gini\Config::Get('app');
